Added Table object factory
A Table object takes a table name and a list.
The list can contain strings and other tables.
Table objects have a get method to get a random item from the list. If it
gets a table, it will get an item off that table, etc., until a string is
found, then return that string.

Instead of a string or table, an object can be included in a list with a
result (string or table), and a weight, which will allow certain options
to have a higher chance of being chosen.